+++ src/vista/entrada/lista.php
@@ -1,23 +1,23 @@
 <?php
 
-/**
- * TODO - Aquí irá la lista de entradas del servidor.
- */
+use dwesgram\modelo\UsuarioBd;
+
 if (!empty($datosParaVista['datos'])) {
     echo "<h3>Esta es la lista de entradas:</h3>";
     echo "<hr>";
-    /*     echo "<div style='display:grid; grid-template-columns:repeat(4,1fr);'>";
- */
+    echo '<div class="grid-container">';
     foreach ($datosParaVista['datos'] as $entrada) {
+        $autor = UsuarioBd::getUsuarioPorId($entrada->getAutor());
 
         echo <<<END
-            <div class="card" style="width: 18rem;">
+            <div class="card grid-item" style="width: 18rem;">
                 <img class="card-img-top" src="{$entrada->getImagen()}" alt="/src/img/alt.webp">
                 <div class="card-body">
+                    <h5 class="card-text">{$autor->getNombre()} escribió</h5>
                     <p class="card-text">{$entrada->getTexto()}</p>
                     <a  href="index.php?controlador=entrada&accion=detalle&id={$entrada->getId()} " class="btn btn-primary">Detalles</a>
         END;
-        if ($sesion->haySesion()) {
+        if ($sesion->haySesion() && $sesion->getId() == $entrada->getAutor()) {
 
             echo "<a href='index.php?controlador=entrada&accion=eliminar&id={$entrada->getId()}' class='btn btn-danger'>Eliminar</a>";
         }
@@ -26,6 +26,7 @@ if (!empty($datosParaVista['datos'])) {
             </div>
         END;
     }
+    echo "</div>";
 } else {
     echo "<h3 class='alert alert-primary'>Tu lista de cosas por hacer está vacía</h3>";
 }
